backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
firebase-admin
email-validator
